Update pending task mechanism again for greater correctness.
On second thought, there were problems in BOTH John's and my prior fixes to
the pending task implementation -- specifically, when a Task was deleted
while it lay on a pending list.  This implementation uses two lists to
solve that problem, without synchronizing the whole router on pending
processing the way John's did.  It also seems a tiny bit faster, probably
because there is no need to reverse the pending list any more.

Not tested nearly enough.