Improve NotifierQueue and Queue behavior on SMP Click.
Jason (Joonwoo) Park reports a desire to use NotifierQueue and Queue on SMP
Click.  This didn't work so well because of notification-related race
conditions.  There was code there to avoid the race conditions, but
commented out, as the locking was observed to make things LOTS slower.
This patch attempts to avoid the race conditions without locking, simply by
double-waking the Notifiers.  There may be bugs.

Also check in comments explicitly noting NotifierQueue/Queue assumptions on
SMP.