Add linuxmodule elements that send packets to/from user level /dev.
The elements are called ToUserDevice and FromUserDevice.  ToUserDevice is a
ring buffer of packets pulled from the rest of the Click config; the user
can extract packets from the buffer via character device 241.
FromUserDevice is a ring buffer of packets that can be pulled from the rest
of the Click config; the user can write packets to the buffer via character
device 240.

The elements are considered experimental, and are only available if you
supply --enable-experimental.

There are several differences relative to Roman's versions.  The wait queue
handling is marginally safer (but there are still problems).  ToUserDevice
stores packets, not packet copies, to minimize copies (not clear this is
faster).  The protocol for interacting with ToUserDevice has changed; now a
read() will return one packet and returns that packet's length.  To read
multiple packets, use ioctl(fd, CLICK_IOC_TOUSERDEVICE_SET_MULTI, 1); this
writes [length, data, (padding to multiple of sizeof(int)), length, data],
etc.

Signed-off-by: Eddie Kohler <kohler@cs.ucla.edu>