More pending race conditions.
Joonwoo Park reports yet another problem with notification on SMP Click.
Try to avoid this problem by only resetting _should_be_scheduled to 0 while
the thread lock is held; this should make reschedule requests stick until
they are satisfied.  This adds an assignment to the fast_unschedule() fast
path, but this will probably not matter as the assignment is near variables
that already needed to be assigned.