Notifiers with different names get different bins.
Required an incompatible change to the Notifier::initialize() API.  The
goal is to encourage Notifiers unlikely to be used together (such as an
EMPTY_NOTIFIER and a FULL_NOTIFIER) to use different notifier words,
leaving Notifiers that WILL be used together a better chance of fitting in
the same word.

At the same time, the number of available notifiers becomes infinite.

Inspired by a list question from Lars Bro.