Add click_random() portable randomness functions.
click_random() returns a random 32-bit integer with at least 31 bits of
randomness.  (CLICK_RAND_MAX is the maximum returned value.)
click_random(a, b) returns a random 32-bit integer between a and b,
inclusive, and does so fairly (the results are uniformly distributed if the
underlying generator is uniformly distributed).

click_random() defaults to random() or rand() at userlevel, if enough
randomness is available, and to the LCRG a=69069, c=5 otherwise.  This LCRG
is also used in the Linux kernel.  The BSD kernel uses the kernel's own
random().

Previously "random()" was the "portable" random number generator, but on
some platform libraries random() gives only 15 bits of randomness.

Use click_random() or click_random(a, b) everywhere random() was used.

Also fix the RandomSeed test element and link glue.o with the tools.