Hotswap: Avoid crashes when dereferencing a Master with associated Routers.
A Router's _state can be ROUTER_LIVE even after it has been disconnected
from its Master -- disconnection affects _running, not _state -- so don't
assume the _master exists.  This is not the normal case, so bugs may
remain, but this fixes special cases like the Router that keeps the
user-level hotswapping Master around.