Atomic operations: Fix swap().
Kevin Springborn reports problems with atomic_uint32_t::swap() in the Linux
kernel.  This was probably due to a bad inline assembly constraint, which
is fixed here.

Also update other code to use Linux symbols when appropriate (hope this
works), and to prefer the standard CONFIG_SMP option to __SMP__.

Also begin the process of removing support for Linux 2.2 kernel modules.