The task lock becomes a Spinlock, not a spinlock_t.
This is necessary for correctness.  Spinlocks are recursive (a thread can
grab the same lock multiple times), a spinlock_t is non recursive (a thread
grabbing a lock multiple times causes deadlock), and we rely on recursive
behavior.

I was hoping someone would test this, to see whether spinlock_ts are
actually much faster than Spinlocks, but it is better to be correct.

Bug reports by Kevin Springborn, analysis and alternate fixes by Joonwoo
Park.