Threading: Differentiate thread blocking from thread locking.
Joonwoo Park points out that Click threads may call blocking functions
and/or schedule().  It is not appropriate for a thread to call blocking
functions and/or schedule() when that thread has prevented preemption via
get_cpu().

The current solution is to separate "locking the task list" from "blocking
task execution."  Clickfs, when executing an exclusive handler, "blocks
task execution."  A thread that wants to examine the task list "locks" the
list.

Tested minimally.