TimedUnqueue: Add new element.
TimedUnqueue is like a Burster with a default BURST of 1.  (Now Burster is
implemented in terms of TimedUnqueue.)

This is a better name for the element.

Also, TimedUnqueue uses a combination of Tasks and notifications to achieve
the goal of imposing relatively little delay on input streams with rates
less than the stated INTERVAL.

Inspired by questions from Øivind Kure about RatedUnqueue CPU usage.