Kernel module allocations: Use GFP_KERNEL if in a file system task.
People have observed handlers failing at times with out-of-memory errors.
This can happen because we pass the GFP_ATOMIC flag to Linux's kmalloc
function, which says "return null if there isn't memory immediately
available."  This is right for normal Click, where god knows what locks are
held; but for the clickfs task, which mostly allocates memory for user
buffers, it's OK to use GFP_KERNEL, which might block until more memory is
available.