FullNoteQueue, NotifierQueue: Improve multithreaded sleepiness handling.
Turning off the empty notifier only on *exactly* achieving the sleepiness
trigger is risky; what if we then notice a race condition and turn *on* the
empty notifier? -- we'd never turn off the notifier again.  Instead, turn
off the notifier whenever we're *above* the sleepiness trigger.  This will
very slightly slow down empty pulls -- but that slowdown is considered
irrelevant.