User-level driver: Refactor select handling for multithread.
Raja reported that the user-level driver, when running with more than one
thread, would take 100% of the CPU whether or not it should.  The problem
was that a thread blocked in select() would still hold the select lock.
This is a serious issue -- if the other thread had called add_select(), the
threads could deadlock, although Raja was seeing a different manifestation
of the problem.  So refactor select handling so that a thread blocked in
select() releases the lock.  Thread safety requires further refactoring the
way we match elements to pollfd structures; it's safer to map elements to
file descriptors.  Now in multithreading, all threads will block when
there's nothing to do, although at most one of those threads will be
blocked on file descriptors.