ARPTable: Fix memory corruption bug caused by nonresponsive ARP requests.
If an ARPQuerier received a duplicate request for an address A where the
previous request for that address had just expired, then
ARPTable::append_query() would touch the entry after freeing it, leading to
leaked packets.

The solution is to treat new requests as keeping an entry just barely
alive.

Reported by Seiichi Tetsukawa.