ControlSocket: Accept TCP PORT arguments like "41930+".
This means start trying to bind to port 41930, then try a bunch more ports
in the neighborhood before giving up.

If PORT ends in a plus, and RETRIES is specified, and a hotswap element
exists (a ControlSocket with the same name and compatible parameters), then
ControlSocket doesn't try multiple ports; it waits in the assumption that
it'll soon take over the hotswapped socket.