click-pretty --dot: Allow the user to set the element label template.
And change the default element label template to include the element name.

Includes documentation, and a more reliable way to find out whether an
element was declared anonymously.  (This should be improved; the
"anonymousness" of an element name is lost by passing through
click-flatten, for example.)

Requested by Bart Braem.