PullSwitch: Add notification support.
Now the empty notification from the currently selected upstream path is
passed downstream.  This means that, for example, in the following
configuration:

   ps :: PullSwitch(1) -> Discard;
   InfiniteSource -> [0] ps;
   Idle -> [1] ps;

the Discard element will go dormant, because it has nothing to pull.  With
the previous PullSwitch, or with SimplePullSwitch, Discard will remain
active and pull like crazy, because it believes the InfiniteSource is
upstream.

You will need to re-configure (to add --enable-simple) and "make elemlist"
for the build to succeed.