Linuxmodule FromDevice: Work in unmodified kernels (--enable-fixincludes).
This code makes use of existing hooks for Ethernet bridging, and will only
work if your kernel was compiled with CONFIG_BRIDGE or
CONFIG_BRIDGE_MODULE.  It replaces the bridge hook with Click code.  When
using Click, it's probably safest to use CONFIG_BRIDGE_MODULE and then not
actually load the module.

Required some refactoring, which simplified the code.

ToDevice also compiles in unmodified kernels, but haven't tested yet if it
runs.