WifiDecap, RadiotapDecap: Fix bugs, add WifiDecap ETHER parameter.
1) WifiDecap does not properly handle cases where padding has been
inserted after the 802.11 header (so as to word-align the next header).

2) After stripping the 802.11 header, WifiDecap always creates and
pushes on an Ethernet header; a configuration parameter should exist to
prevent this from occurring if so desired.

3) RadiotapDecap does not correctly read the "frame includes FCS" flag
from radiotap headers, meaning packets are not stripped of their FCS (at
the end of the packet) as intended.

4) RadiotapDecap does not re-set the mac-header pointer
(p->set_mac_header) after stripping the radiotap header.