Handlers: Separate comprehensive handlers for read & write.
This is an INCOMPATIBLE CHANGE, although I expect most user code to
continue to work without changes.

And take this opportunity to improve constant names and remove deprecated
functions.

After this change, a handler can have two different "comprehensive" (i.e.,
set_handler) handlers, one for read and one for write.  This will be useful
for data_handlers.

A side effect is that the Handler::user_data1() and Hander::user_data2()
values might be overwritten.  For example, after this:

  set_handler("foo", Handler::h_read, func1, 0, 1);

when func1 is called, it will see h->read_user_data() == 0 and
h->write_user_data() == 1.  But if this is also called:

  set_handler("foo", Handler::h_write, func2, 2, 3);

then when either func1 or func2 is called, they will see
h->read_user_data() == 0 and h->write_user_data() == 3 -- the read
handler's write_user_data is overwritten, as is the write handler's
read_user_data.