User level: Fix signal handling.
Previously, we attempted to make signal delivery reliable by blocking a
signal as soon as it was delivered.  This does not work, because changes to
a process's signal mask from within a signal handler are ignored.  So
instead leave the Click signal handler in place until the signal is about
to be handled by the Master.  THAT is the right time to block the signal,
and the right time to reset the signal handler to the default.