User-level multithreading: Fix race condition with task rescheduling.
If thread A wakes up a task just as that task's owning thread, B, goes to
sleep, the following code can run simultaneously:

A                               B
-----------------------------   ------------------------------
mark thread pending;            mark thread blocked;
if (thread blocked)             if (anything pending)
   write byte to thread pipe;      don't sleep;

We need barriers after each "mark" and before the corresponding test.  But
we only had a barrier on B.  This adds a barrier to both sides.

This fixes a problem reported by Beyers Cronje.  Thanks to Cliff Frey for
discussion.