Timer heap: Improve cache behavior by duplicating expiry.
Previously, the Timer heap looked like an array of Timer pointers.
That meant that walking the heap required an indirect load for every
comparison, since the expiry was stored with the Timer, not in the
array.  Bad cache behavior.

This checkin duplicates a scheduled Timer's expiry in the heap array.
As a result, heap-walking comparisons use data already in cache.  (The
place() functionality, used to update a timer's current heap position,
still requires an indirect write.)  This makes TimerTest(BENCHMARK
500000) run about 40% faster on my modern Intel imac.  Before: ~1.77s
real time, after: ~1.01s real time.