UDPTimeCheck - element for UDP e2e delay estimation
UDPTimeCheck element can be handy when computing end-to-end UDP packet
delays.  The element embeds a timestamp and a sequence number into a packet and
adjusts the checksum of the UDP packet.  Once the initial timestamp has been
placed into the payload of the UDP packet, a time difference can be computed
once a packet passes through another UDPTimeCheck element.  The data can be
accessed by examining the packet payload (e.g., do a tcpdump and then do post
processing of the data).  The element uses partial checksums to speed up the
processing.

Signed-off-by: Roman Chertov