StoreUDPTimeSeqRecord - UDP element for measuring e2e delay
StoreUDPTimeSeqRecord element can be handy when computing end-to-end UDP packet
delays.  The element embeds a timestamp and a sequence number into a packet and
adjusts the checksum of the UDP packet.  Once the initial timestamp has been
placed into the payload of the UDP packet, a time difference can be computed
once a packet passes through another StoreUDPTimeSeqRecord element.  The data can be
accessed by examining the packet payload (e.g., do a tcpdump and then do post
processing of the data).  The element uses partial checksums to speed up the
processing.  The element works with IPv4 and IPv6 packets.  In case of IPv6,
it assumes that the next header field is UDP.  If this is not the case, the
element will discard the packet.

Signed-off-by: Roman Chertov <roman.o.chertov@aero.org>