Make DSRRouteTable not mess with incoming packets' TTLs
DSRRouteTable::add_dsr_header() was setting the TTL of
IP packets it should send out to 255.

Leaving the TTL at whatever the incoming IP packet provides (usually
the operating system's default) is a IMHO saner option than stubbornly
setting it to 255.

When forwarding, the TTL can be decremented easily using the DecIPTTL
element.

Some traceroute, tracepath, or my IPMP tools work much better when
the TTL is decremented at each hop, like it is supposed to.

Signed-off-by: Florian Sesser <sesser@in.tum.de>