Change (BW)RatedUnqueue and (BW)RatedSplitter elements to use TokenBuckets.
The biggest functional change here is that these elements now have
consistent and customizable burst behavior.  The elements now all use
token buckets configured by default with a capacity of 10 milliseconds
worth of traffic.  This leads to a slight spike in traffic after
periods of idle time.  The capacity of the token bucket (and therefor
the maximum burst size) can be configured using an absolute size
(BURST_SIZE/BURST_DURATION) or by specifying a duration of time
(BURST_DURATION).

I believe (untested) that GapRate would also have bursty behavior,
except for the special case of the very first packet through it.  At
that point there was no burst.  After idle periods in the future,
GapRate would provide some amount of bursting (I believe on the order
of 1 second, but am not sure).

The other big change is that these elements all now include timers
which they are willing to use, so they should not ever end up pegging
the CPU.

This also greatly improves performance.  I believe that this is
because of TokenBucket's relative efficiency to GapRate.
  cat myconfig
    InfiniteSource(LIMIT 4000000, TIMESTAMP false, STOP true)
    -> Queue(CAPACITY 100)
    -> RatedUnqueue(4000000000)
    -> Discard
  time ./userlevel/click myconfig
    real		      0m6.319s
    user		      0m3.212s
    sys		      0m3.088s
  time ./userlevel/click myconfig
    real		      0m1.729s
    user		      0m1.200s
    sys		      0m0.520s

Signed-off-by: Cliff Frey <cliff@meraki.com>
Reviewed-by: Eddie Kohler <ekohler@gmail.com>