Connections: Support parallel connections "a [0,1,2] -> b, c, d".
This means a[0] -> b; a[1] -> c; a[2] -> d (a change from previously, which
would have introduced all-pairs connections).

The rules:
* If at least one side of a connection specifies exactly one port, then
  create all connections to that port.
* Otherwise, both sides of the connection specify more than one port.  The
  number of ports on each side must match (call it N), and exactly N
  connections are created, port to port.

Also support expandable port specifications, such as "a[]".  This is the
same as "a[0,1,2,3,...]", with as many additional ports as necessary to
match the number of inputs.  "a[1,]" is also expandable; it means
"a[1,2,3,4,...]".  Any connection attempt can involve at most one
expandable port.