Language: Interpolate "input" or "output" within compounds.
Specifically, "{ -> }" now means the same as "{ input [0] -> [0] output }".
So does "{ [0] -> [0] }".

This syntax has admirable concision.  Unfortunately, it also makes
some semicolons mandatory (previously they could always be left off).
Consider "{ Idle -> [0] ; Idle }" and "{ Idle -> [0] Idle }".  The
non-interpolated syntax takes precedence (i.e., prefer shift to
reduce).

The rules: At the beginning of a chain "input" is assumed.  At the end
of a chain "output" is assumed.  Inference requires port numbers,
except that "[0]" is inferred if no ports at all are given.  In other
words, "{ a, -> b }" is an error, and so is "{ [0], , [1] -> [0] }",
but "{ -> b }" is allowed (input[0] is inferred).

Signed-off-by: Eddie Kohler <ekohler@gmail.com>