Language: Add element groups.
An element group is like a compound element that has the same scope as its
surrounding environment.  Element groups take no arguments and do not
support overloading.  Element group classes don't make sense.  Groups are
written with parentheses; these three lines have identical meaning:

   a :: Idle -> b :: Idle;
   a :: Idle -> ( -> b :: Idle );
   a, b :: Idle; a -> (->b);

Note that unlike compounds ("{ -> b :: Idle }" or "{->b}"), the "b" element
is part of the enclosing scope.

Element groups are intended for use with the new comma syntax.  For
example, compare this:

   c :: Classifier(..., ...);
   post_classification :: ...;
   c [0] -> [1] analyzer [1] -> post_classification;
   c [1] -> c2 :: Classifier(..., ...);
            c2 [0] -> post_classification;
            c2 [1] -> Paint(1) -> post_classification;

with this:

   c :: Classifier(..., ...) [] ->
     [1] analyzer [1],
     (-> c2 :: Classifier(..., ...) [] -> (->), Paint(1) ->)
   -> post_classification :: ...;

Signed-off-by: Eddie Kohler <ekohler@gmail.com>