Language: Require '=>' for many-to-many connections.
It is a syntax error to use '->' for a many-to-many connection.

Also, '=>' will sometimes interpret a missing port as an expandable
port: if one side of a many-to-many connection has one element with
port specification (and the other side has >1 port), then the missing
port is treated as '[]'.  The following three lines have equivalent
effect:

   a [0,1,2] => b, c, d;
   a [] => b, c, d;
   a => b, c, d;

So do these three lines:

   a [0,1] => [3,4] b;
   a [] => [3,4] b;
   a => [3,4] b;

These are errors, though, because the interpreted expandable port
creates more than one expandable port total:

   a => [3,] b;
   a => [] b;

Signed-off-by: Eddie Kohler <ekohler@gmail.com>