Language: Fix use of element groups inside compound elements.
This worked with the tool lexers, but not with the driver lexers.  The
problem was that tunnels, such as those used to implement element group
inputs and outputs, were unscoped.  This is OK for compound elements, which
create the tunnels only at expansion time, but not for element groups.  Fix
through a little trickery and add a test.  Cliff Frey report, thanks!

Signed-off-by: Eddie Kohler <ekohler@gmail.com>