Linuxmodule: Install Click kernel symbol CRC file.
And refer to it when building a package.

Unfortunately this is not good enough in practice, because Linux only
supports symbols that are 60 characters long (on 32-bit machines; on
64-bit machines the max length is 56 characters), and some Click symbols,
being C++ mangled, are longer.

It looks like, if you want to use Click packages on patchless, you need
to use a kernel configured without modversions.