Args: type-safe argument parsing class.
The Args class is cleaner, easier to use, and safer than previous
varargs-based argument parsing.  It's safer because it's type safe, and
because it will not use global variables like cp_errno or the underlying
cp_values array.

This has been long coming -- years.  Much thanks to Cliff Frey for
discussions and suggestions.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>