classification.cc: fix bug in DominatorOptimizer
Without this change, this config will fail the assertion at the start of
DominatorOptimizer::last_common_state_in_lists.

Idle
-> ipf :: IPFilter(allow tcp && dst 10.0.0.0/32 && src 2.0.0.0/32,
                   allow tcp && dst 10.0.0.0/32 && dst port 80 && src 2.0.0.0/32,
                   allow all)
-> Discard;

I can confess that I do not actually understand the code involved enough well
enough to know if this is the best fix, however it seems like it is safe to me.
I am also unsure of a good test case to add.  I can add the above, but it is
incredibly fragile... reordering or removing anything in it seems to be enough
to make it stop failing.