Args keywords must begin with a letter or underscore.
[0-9.:?!] are only accepted starting with the second character.

I've wanted to do this anyway -- keywords that start with digits are just
weird -- but it will also help migration from cp_va in a corner case.

When cp_va parses an optional positional argument, it uses the positional
string as long as it matches no parsed keyword.  For example, given the
config string "22 2", this:

  String a, b, c;
  cp_va_kparse(conf, this, errh,
               "A", cpkP, cpString, &a,
               "B", 0, cpString, &b,
               "C", 0, cpString, &c, cpEnd);

would set a="22 2": although "22" was a valid keyword, it didn't match "B"
or "C".

Unfortunately, when Args parses an argument, it does not have the set of
all known parse keywords.  Therefore, Args uses a positional string only if
it COULD NOT set ANY keyword.  The Args version of the above would
previously not set "A".

However, after this commit, the "22 2" example will work, since "22" is not
a valid keyword.  Hopefully this will fix common cases without causing too
much incompatibility itself.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>