IPReassembler: add features and debugging info
IPReassembler can now be used on packets with ethernet headers, in which
case the ethernet header of fragment zero is preserved.

It also now supports the MTU_ANNO which stores the size of the largest
fragment.  This is useful if you want to reassemble and then refragment
packets without increasing the MTU.

I believe that this change also fixes a double-free by removing a p->kill()

Signed-off-by: Cliff Frey <cliff@meraki.com>