Give each thread its own SelectSet.
Similar reasons and benefits as the TimerSet version.

Simultaneously simplify the wakeup plan for signals and cross-thread
communication.  The _wake_pipe, previously multithread-only, moves to
SelectSet, and is used even when single threaded.  This requires care (an
optimized path for the no-selector case must be aware of the _wake_pipe).

This slows the NullTask test down relative to the previous commit,
something to be addressed later.

And make sure that the select lock is released whenever signals are
processed.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>