Multithread: Use compare_swap on Notifiers.
This appears to fix the bug with InfiniteSource->Queue->Discard, IS and D
on different threads, mentioned earlier.  Which makes some sense, although
I have not fully worked out an example.  There are disadvantages.  Now
every wake() or sleep() involves a compare_swap, which may be more
expensive than the previous click_fence()+atomic_or() pair.  But this is
unmeasured.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>