AddressInfo: Support ":gw", and read network prefixes for Ethernet devices.
Given an IP prefix "X", the address "X:gw" is the default gateway address
for X -- namely, the network address for X, with the lowest-order bit
turned on.

Given an Ethernet device "X", on Linux and in the Linux kernel module,
"X:net" is now the default IPv4 network address.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>