Classification optimization: Fix an optimization bug.
The bug could cause a test like "tcp or udp" to completely dominate the
result of a test like "tcp and tcp and tcp".  The problem was we set a
variable to -1 to mean "this test has had conflicting results on
predecessor paths", but then could later increment that -1 to 0 (meaning
"this test has no known result") and then 1 (meaning "this test has a known
result").

Signed-off-by: Eddie Kohler <ekohler@gmail.com>