TokenBucket: Rename interface elements.
"Epoch" means the beginning point of time, not a unit of time; so
"epochs_until_contains", for example, is a terrible name.  Base the new
names on the upcoming C++ standard's classes and names.  However, we need a
new type, that of unsigned differences between time points.  This isn't a
"duration".  Is it an "unsigned_duration_type"?  Currently using
"ticks_type", plus "time_until" verbiage.  (This isn't crazy, since
functions like "time_since_epoch" in the standard return duration_type.)

Signed-off-by: Eddie Kohler <ekohler@gmail.com>