add_*_handler*: use const char* instead of String in many places
Specifically this adds a add_(read|write)_handler overload that takes
const char * for the name, and it changes add_data_handlers to only
accept const char *.

This saves on code size, because every add_handlers() was doing a string
construct/deconstruct on every single handler added.  This shrinks the stripped
click executable from 2320608 to 2271456 bytes (2.1%).  The space savings is
much higher (above 4%) on MIPS machines, where the function call and relocation
overhead is much higher.

In addition, this saves a small amount of RAM because the strings for handler
names themselves are now stored with String::make_stable instead of copying the
actual bytes.  On a large click configuration with many different elements, this
can save ~10kB.

Signed-off-by: Cliff Frey <cliff@meraki.com>