Timestamp: Add recent(), now_steady(), and recent_steady().
The steady() functions return the current time as measured by a steady
clock: a monotonic clock that never moves backwards.  We call the clock
"steady" rather than "monotonic" or "up" because C++0x calls this clock the
steady_clock, and that's a good name.

The recent() functions return a cached copy of a recent now() or
now_steady().  At userlevel at the moment, these functions actually just
call now() or now_steady().  In the kernels, though, they will be faster
than now() or now_steady().

Signed-off-by: Eddie Kohler <ekohler@gmail.com>