RouterThread: A please_stop_driver() immediately stops most tasks.
Checking the "_stop_flag" before every task is actually pretty expensive on
a null-task list.  So is storing the number of tasks to check in
RouterThread (by, say, 5% or so).  So, instead, shift the expense to the
stop case.  When asked to stop the driver, mark all tasks as unrunnable and
put them on the pending list.  This will, in most cases, cause the router
to stop processing tasks immediately, rather than running tasks a couple
more times before processing the stop request; and it is very close to as
fast as previously.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>