Deque: Rewrite template for fuller functionality and to use array_memory.
But getting all the cases right for best possible array_memory use --
i.e., at most 3 memmoves per operation on trivially copyable types --
is way too much work and probably not worth it. So don't bother.

This adds one important function: it is now OK to call, e.g.,
deque.insert(deque.begin(), deque[1]), which was not OK before (because
the memory for the deque[1] reference might disappear).

Also, document Deque, and rename it to Deque from DEQueue (the new name
is what the standard uses).

Signed-off-by: Eddie Kohler <ekohler@gmail.com>