script: add catq/printq/printnq handlers
These allow binary-safe handlercalls.

Specifically, if there is a file containing " foo" that should be passed as an
argument to a handler, and output the results of the handler to a file, it was
impossible before this change (the space at the beginning would be eaten).

now this can be run:
  printnq >output_file my_elt.my_handler $(catq input_file)