PullSwitch: Add precise wake notification.
Now, in a situation like this,

  src0, src1, src2, src3, src4 => p :: PullSwitch(0) -> sink;

"sink" will only be scheduled if "src0" becomes nonempty. Previously,
"sink" would be scheduled if any of src1-4 become nonempty; it would then
go back to sleep, but still.

Cliff Frey and Justin Delegard reported an issue, which they solved in a
different way.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>