click: Make preemption friendly
This change should effectively be a NOP for when CONFIG_PREEMPT is not
set, but is an attempt to make Click work better in the case of
CONFIG_PREEMPT being set.

In my testing, I was able to run with moderate activity for hours,
as well as unload and reload the system click configuration repeatedly.

This is good enough for running with CONFIG_LOCKDEP and CONFIG_PREEMPT
both enabled, which tends to make race conditions come out of hiding.

Specific changes:

    sync.hh: Move to using a mutex so that timer handlers
    and other things can sleep. This was a problem for
    elements which use mutexes for their synchronization.

    clickfs: Switch from semaphore to mutex

    Switch clickfs locking from using semaphores to mutex, as
    write handlers can now use mutexes for their synchronization.

    proclikefs: Switch some semaphores to be mutexes

    Some of the file system is protected by semaphores, so
    change only the appropriate ones.

    master: Make the master lock a mutex

    By making the master lock a mutex, sleeping functions can
    be safely called throughout click.

    Remove the use of preempt_disable(), added in
    e2e08269d8eb1bcef5b54b95738a39c5e73f0890.
    It appears that a0717e1fdad15ab3f29a76b204ecb710477c0372
    may have eliminated the underlying issue.