ToHost: clear skb->sp on packets going to linux
Linux doesn't clear the sec_path variable in skb when the packet gets
sent to click.  If this packet originally came from linux and had xfrm
states applied, then they would still be saved in this member.  When
linux receives the packet again it will require that an appropriate
policy exists to match the packet and these xfrm states.

Clearing the sec_path makes linux treat the packet as a new packet
coming in off the wire.