FromHost: add a BURST option similar to FromDevice
If traffic goes through a FromDevice, to a ToHost, and then back
through a FromHost, then when run_task is round robining between
FromDevice and FromHost, the FromDevice will get to send multiple
packets for every single packet that FromHost sends.  Adding a BURST
option to FromHost allows management of the load between the two
elements.