click: rewrite clickfs handler_strings
Click uses a pool of Strings to communicate with handlers.  The handler code
actually runs in the context of the reading process and the handler_strings
code was not thread-safe.  The primary issue is that the String[] could be
resized while there were outstanding references to the old location.

This change pushes a String and it's flags into a HandlerString object,
holds pointers to unused HandlerStrings in a Deque.

Reviewed-By: cliff@meraki.com