linuxmodule: fix another HandlerStrings issue
In the new alloc_handler_string() code, we no longer have access to Handler*
so the caller is responsible to setting HandlerString::flags.

Unfortunately, this member also indicates whether the HS is in-use or not
since in-use and not-in-use HSs exist in the same List.  Since it is the
callers responsibility to set flags to an actual value, it is possible for
one caller to mistakenly take an in-use HandlerString before a previous
caller has written flags.