linuxmodule/clickfs: change read locks to write locks (temporary hack)
In ClickIno::prepare, we might invoke ClickIno::grow, which potentially
reallocates and moves a shared structure.  This is not safe under a read
lock since other readers might be holding references to objects in the
modified array.  Unfortunately, there are a bunch of places in
linuxmodule/clickfs that invoke ClickIno::prepare under a read lock instead
of a write lock.

This commit is a temporary fix until either ClickIno or linuxmodule/clickfs
can be fixed properly.